add mocked last online statistic to host
